requirements.txt
boto3==1.18.51
botocore==1.21.51
bs4==0.0.1
Distance==0.1.3
django-background-tasks==1.2.5
dj-database-url==0.5.0
Django==3.2.7
django-storages==1.11.1
docopt==0.6.2
gunicorn==20.1.0
psutil==5.8.0
psycopg2-binary==2.9.1
django-materializecss-form==.1.17